Refactor: local provider공급 변경
- 분산되어 공급되던 composition local을 scaffold내부에서 공급하도록 변경
