+++ core/ui/src/main/java/com/kolown/porring/core/ui/compositionlocal/LocalPaddingValue.kt
@@ -1,0 +1,8 @@
+package com.kolown.porring.core.ui.compositionlocal
+
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.runtime.compositionLocalOf
+
+val LocalPaddingValues = compositionLocalOf<PaddingValues> {
+    error("No LocalPaddingValues provided")
+}
